refactor: move config dataio into assets dataio
