+++ src/student_success_tool/dataio/assets.py
@@ -1,9 +1,23 @@
 import logging
 import pathlib
+import typing as t
+
+import pydantic as pyd
 
 from . import read
 
 LOGGER = logging.getLogger(__name__)
+
+S = t.TypeVar("S", bound=pyd.BaseModel)
+
+
+def read_config(file_path: str, *, schema: type[S]) -> S:
+    """
+    Read config from ``file_path`` and validate it using ``schema`` ,
+    returning an instance with parameters accessible by attribute.
+    """
+    cfg = read.from_toml_file(file_path)
+    return schema.model_validate(cfg)
 
 
 def read_features_table(file_path: str) -> dict[str, dict[str, str]]:
